Use hash in filename as the asset version numbner
Previously, asset-loader was returning `null` as the asset version
number for assets which contained hashes in their filename. This was
done under the assumption that filename-based cache busting meant that
there was no need to implement any other type of cache-busting.

However, there are still cricumstances where registering a script with a
version number is desirable. One example is when using an nginx http
concatenation setup similar to WordPress VIP - if the version number
doesn't increment, the static concatenated file will be missing some of
it's components.